requirements.txt
flask
pandas
numpy
scikit-learn
xgboost
mysql-connector-python